Rework Mui2 UIs to fit new master branch
