+++ src/main/java/gregtech/common/tileentities/machines/multi/nanochip/modules/Splitter.java
@@ -1,8 +1,6 @@
 package gregtech.common.tileentities.machines.multi.nanochip.modules;
 
 import static com.gtnewhorizon.structurelib.structure.StructureUtility.ofBlock;
-import static gregtech.api.modularui2.GTGuis.createPopUpPanel;
-import static gregtech.api.modularui2.GTGuis.mteTemplatePanelBuilder;
 import static gregtech.common.tileentities.machines.multi.nanochip.MTENanochipAssemblyComplex.NAC_MODULE;
 
 import java.util.ArrayList;
@@ -11,48 +9,27 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.nbt.NBTTagList;
 import net.minecraft.nbt.NBTTagString;
-import net.minecraft.network.PacketBuffer;
 import net.minecraft.util.EnumChatFormatting;
 
 import org.jetbrains.annotations.NotNull;
 
-import com.cleanroommc.modularui.api.IPanelHandler;
-import com.cleanroommc.modularui.api.drawable.IIcon;
-import com.cleanroommc.modularui.api.drawable.IKey;
-import com.cleanroommc.modularui.api.widget.IWidget;
-import com.cleanroommc.modularui.drawable.GuiTextures;
-import com.cleanroommc.modularui.factory.PosGuiData;
-import com.cleanroommc.modularui.screen.ModularPanel;
-import com.cleanroommc.modularui.screen.RichTooltip;
-import com.cleanroommc.modularui.utils.Alignment;
-import com.cleanroommc.modularui.utils.serialization.IByteBufAdapter;
-import com.cleanroommc.modularui.value.sync.GenericSyncValue;
-import com.cleanroommc.modularui.value.sync.PanelSyncManager;
-import com.cleanroommc.modularui.widget.ParentWidget;
-import com.cleanroommc.modularui.widget.WidgetTree;
-import com.cleanroommc.modularui.widgets.ButtonWidget;
-import com.cleanroommc.modularui.widgets.ListWidget;
-import com.google.common.collect.ImmutableList;
 import com.gtnewhorizon.structurelib.structure.IStructureDefinition;
 import com.gtnewhorizon.structurelib.structure.ISurvivalBuildEnvironment;
 
 import gregtech.api.GregTechAPI;
-import gregtech.api.enums.Dyes;
 import gregtech.api.interfaces.metatileentity.IMetaTileEntity;
 import gregtech.api.interfaces.tileentity.IGregTechTileEntity;
-import gregtech.api.modularui2.GTGuiTextures;
+import gregtech.api.metatileentity.implementations.gui.MTEMultiBlockBaseGui;
 import gregtech.api.recipe.check.CheckRecipeResult;
 import gregtech.api.recipe.check.CheckRecipeResultRegistry;
 import gregtech.api.util.GTUtility;
 import gregtech.api.util.MultiblockTooltipBuilder;
-import gregtech.common.modularui2.widget.ColorGridWidget;
+import gregtech.common.tileentities.machines.multi.gui.nanochip.SplitterGui;
 import gregtech.common.tileentities.machines.multi.nanochip.MTENanochipAssemblyModuleBase;
 import gregtech.common.tileentities.machines.multi.nanochip.hatches.MTEHatchVacuumConveyorInput;
 import gregtech.common.tileentities.machines.multi.nanochip.hatches.MTEHatchVacuumConveyorOutput;
@@ -352,243 +329,8 @@ public class Splitter extends MTENanochipAssemblyModuleBase<Splitter> {
     }
 
     @Override
-    public ModularPanel buildUI(PosGuiData data, PanelSyncManager syncManager) {
-        ModularPanel ui = mteTemplatePanelBuilder(this, data, syncManager).build();
-        IPanelHandler popupPanel = syncManager.panel("popup", (m, h) -> createRuleManagerPanel(syncManager), true);
-        GenericSyncValue<Map<Integer, ColorRule>> listSyncer = new GenericSyncValue<>(
-            () -> colorMap,
-            map -> { colorMap = map; },
-            new ColorMapAdapter());
-        syncManager.syncValue("rules", 0, listSyncer);
-
-        return ui.child(new ButtonWidget<>().onMousePressed(mouseButton -> {
-            popupPanel.openPanel();
-            return popupPanel.isPanelOpen();
-        })
-            .background(GTGuiTextures.BUTTON_STANDARD)
-            .tooltip(tooltip -> tooltip.add("Add Rule"))
-            .pos(153, 5)
-            .size(18, 18));
-    }
-
-    public ModularPanel createRuleManagerPanel(PanelSyncManager syncManager) {
-        ModularPanel ui = createPopUpPanel("gt:splitter:rules_manager", false, false);
-
-        ListWidget<IWidget, ?> list = new ListWidget<>();
-        list.childSeparator(IIcon.EMPTY_2PX);
-        list.size(168, 138);
-        list.pos(4, 21);
-
-        // Add existing rules
-        for (Map.Entry<Integer, ColorRule> entry : colorMap.entrySet()) {
-            int id = entry.getKey();
-            ColorRule rule = entry.getValue();
-            if (rule == null) continue;
-            Byte input = rule.getInputColor();
-            List<Byte> outputs = rule.getOutputColors();
-            // spotless:off
-            list.child(createColorManager(syncManager, Stream.of(input).collect(Collectors.toList()), outputs, id));
-            // spotless:on
-        }
-
-        return ui.child(list)
-            .child(new ButtonWidget<>().onMousePressed(mouseButton -> {
-                list.child(createColorManager(syncManager, null, null, null));
-                WidgetTree.resize(ui);
-                return true;
-            })
-                .pos(80, 4)
-                .size(16, 16)
-                .overlay(GuiTextures.ADD)
-                .tooltip(tooltip -> tooltip.add("Add new Rule")))
-            .posRel(0.75F, 0.5F);
-    }
-
-    public IWidget createColorManager(PanelSyncManager syncManager, List<Byte> inputSelected, List<Byte> outputSelected,
-        Integer indexOverride) {
-        ColorGridWidget inputGrid = new ColorGridWidget();
-        ColorGridWidget outputGrid = new ColorGridWidget();
-        ColorGridSelector selector = new ColorGridSelector(syncManager, indexOverride);
-
-        return selector
-            // Arrow icon
-            .child(
-                GTGuiTextures.PROGRESSBAR_ARROW_STANDARD.getSubArea(0F, 0F, 1F, 0.5F)
-                    .asWidget()
-                    .size(20, 18)
-                    .posRel(0.5F, 0.5F))
-            .setInputGrid(
-                (ColorGridWidget) inputGrid.setInitialSelected(inputSelected)
-                    .setMaxSelected(1)
-                    .build()
-                    .pos(5, 17))
-            .setOutputGrid(
-                (ColorGridWidget) outputGrid.setInitialSelected(outputSelected)
-                    .setMaxSelected(16)
-                    .build()
-                    .pos(121, 17))
-            // Input grid color display
-            .child(
-                IKey.dynamic(() -> inputGrid.getName(0))
-                    .asWidget()
-                    .scale(0.8F)
-                    .alignment(Alignment.Center)
-                    .size(42, 8)
-                    .pos(4, 5))
-            // Output grid color display
-            .child(
-                // spotless makes this look vile and disgusting and abominable and atrocious and yucky and horrid and
-                // offensive to the eyes and nasty and foul and repugnant and abhorrent and deplorable and nauseating
-                // and Dirty
-                // spotless:off
-                IKey.dynamic(() -> switch (outputGrid.getAmountSelected()) {
-                        case 0: yield "None";
-                        case 1: yield outputGrid.getName(0);
-                        default: yield "[Hover]";
-                    })
-                // spotless:on
-                    .asWidget()
-                    .tooltipAutoUpdate(true)
-                    .tooltipBuilder(t -> getInfo(t, outputGrid))
-                    .scale(0.8F)
-                    .alignment(Alignment.Center)
-                    .size(42, 8)
-                    .pos(120, 5))
-            // Delete button
-            .child(
-                new ButtonWidget<>().tooltip(t -> t.add("Delete"))
-                    .onMousePressed(a -> {
-                        selector.removeColorData();
-                        WidgetTree.resize(selector.getParent());
-                        return true;
-                    })
-                    .overlay(GTGuiTextures.OVERLAY_BUTTON_CROSS)
-                    .pos(80, 5)
-                    .size(8, 8))
-            .size(166, 58)
-            .background(GTGuiTextures.BACKGROUND_POPUP_STANDARD);
-    }
-
-    public void getInfo(RichTooltip t, ColorGridWidget grid) {
-        List<Byte> selected = grid.getSelected();
-        int amount = selected.size();
-        if (amount < 2) return;
-        t.pos(RichTooltip.Pos.ABOVE)
-            .add("Currently selected:\n");
-        for (int i = 0; i < amount; i++) {
-            boolean shouldNewLine = (((i - 2) % 3) == 0);
-            Dyes color = Dyes.get(selected.get(i));
-            String name = color.getLocalizedDyeName();
-            if (color == Dyes.dyeBlack) color = Dyes.dyeGray;
-            t.add(color.formatting + name + IKey.RESET + (shouldNewLine ? "\n" : ", "));
-        }
-    }
-
-    private class ColorGridSelector extends ParentWidget<ColorGridSelector> {
-
-        int id;
-        ColorGridWidget inputGrid;
-        ColorGridWidget outputGrid;
-        PanelSyncManager manager;
-        GenericSyncValue<Map<Integer, ColorRule>> colorMapSyncer;
-
-        public ColorGridSelector(PanelSyncManager syncManager, Integer indexOverride) {
-            super();
-            manager = syncManager;
-            colorMapSyncer = (GenericSyncValue<Map<Integer, ColorRule>>) syncManager.getSyncHandler("rules:0");
-            if (indexOverride == null) {
-                while (colorMap.get(id) != null) {
-                    id++;
-                }
-            } else id = indexOverride;
-        }
-
-        @Override
-        public void onInit() {
-            super.onInit();
-            saveColorData();
-        }
-
-        public void removeSelector() {
-            IWidget widget = getParent();
-            if (!(widget instanceof ListWidget list)) return;
-            list.getChildren()
-                .remove(this);
-        }
-
-        public ColorGridSelector setInputGrid(ColorGridWidget widget) {
-            inputGrid = widget.onButtonToggled(this::saveColorData);
-            return this.child(inputGrid);
-        }
-
-        public ColorGridSelector setOutputGrid(ColorGridWidget widget) {
-            outputGrid = widget.onButtonToggled(this::saveColorData);
-            return this.child(outputGrid);
-        }
-
-        private void saveColorData() {
-            if (manager.isClient()) {
-                colorMap.put(id, thisAsRule());
-                colorMapSyncer.setValue(colorMap);
-            }
-        }
-
-        private void removeColorData() {
-            if (manager.isClient()) {
-                colorMap.remove(id);
-                colorMapSyncer.setValue(colorMap);
-            }
-            removeSelector();
-        }
-
-        public ColorRule thisAsRule() {
-            List<Byte> input = inputGrid.getSelected();
-            List<Byte> output = outputGrid.getSelected();
-            if (input.isEmpty()) input = ImmutableList.of((byte) -1);
-            if (output.isEmpty()) output = ImmutableList.of((byte) -1);
-            return new ColorRule(input.get(0), output);
-        }
-    }
-
-    private static class ColorMapAdapter implements IByteBufAdapter<Map<Integer, ColorRule>> {
-
-        @Override
-        public Map<Integer, ColorRule> deserialize(PacketBuffer buffer) {
-            Map<Integer, ColorRule> list = new HashMap<>();
-            int size = buffer.readInt();
-            for (int i = 0; i < size; i++) {
-                int id = buffer.readInt();
-                Byte input = buffer.readByte();
-                List<Byte> outputs = new ArrayList<>();
-                int ruleCount = buffer.readInt();
-                for (int j = 0; j < ruleCount; j++) {
-                    outputs.add(buffer.readByte());
-                }
-                list.put(id, new ColorRule(input, outputs));
-            }
-            return list;
-        }
-
-        @Override
-        public void serialize(PacketBuffer buffer, Map<Integer, ColorRule> map) {
-            buffer.writeInt(map.size());
-            for (Map.Entry<Integer, ColorRule> entry : map.entrySet()) {
-                ColorRule rule = entry.getValue();
-                Byte input = rule.getInputColor();
-                List<Byte> outputs = rule.getOutputColors();
-                buffer.writeInt(entry.getKey());
-                buffer.writeByte(input);
-                buffer.writeInt(outputs.size());
-                for (Byte dye : outputs) {
-                    buffer.writeByte(dye);
-                }
-            }
-        }
-
-        @Override
-        public boolean areEqual(Map<Integer, ColorRule> t1, Map<Integer, ColorRule> t2) {
-            return t1.equals(t2);
-        }
+    protected @NotNull MTEMultiBlockBaseGui getGui() {
+        return new SplitterGui(this);
     }
 
     public static class ColorRule {
